Changes made based on review comments
1)updated depends on method for forgotPassword on changepassword & megered both test cases into single class
2)updated verificationToastMessage method - to fetch the first locator
3)minor changes
-removing () in signupTest
- camel case changes
4)ingegrated fetchproperty and verifyToastMessage in forgotPassword test case
