+++ UISanitySuite.xml
@@ -1,5 +1,5 @@
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
-<suite name="SampleSuite">
+<suite name="SanitySuite">
     <test name="Test">
         <classes>
 
@@ -18,6 +18,13 @@
             <class name="org.shikshalokam.frontend.TestBecomeAMentor">
                 <methods>
                     <include name="becomeMentor"/>
+                </methods>
+            </class>
+
+            <class name="org.shikshalokam.frontend.TestChangeAndForgotPassword">
+                <methods>
+                    <include name="changePassword"/>
+                    <include name="forgotPassword"/>
                 </methods>
             </class>
