+++ src/test/java/org/shikshalokam/uiPageObjects/AppAllPages.java
@@ -7,11 +7,11 @@ public class AppAllPages {
 
     private static final Logger logger = LogManager.getLogger(AppAllPages.class);
     public static AppLoginPage loginPage = new AppLoginPage("MentorED");
-    public static AppSignupPage SignupPage = new AppSignupPage("MentorED");
+    public static AppSignupPage signupPage = new AppSignupPage("MentorED");
     public static AppBecomeMentorPage mentorPage =new AppBecomeMentorPage("MentorED");
     public static AppWelcomePage welcomePage =new AppWelcomePage("MentorED");
     public static AppWorkspacePage workspacePage =new AppWorkspacePage("MentorED");
-    public static AppProfilePage ProfileDetailsPage =new AppProfilePage("MentorED");
+    public static AppProfilePage profileDetailsPage =new AppProfilePage("MentorED");
     public static AppChangePasswordPage changePasswordPage =new AppChangePasswordPage("MentorED");
     public static AppResetPasswordPage resetPasswordPage =new AppResetPasswordPage("MentorED");
